support fp16

+++ gluoncv/model_zoo/rpn/proposal.py
@@ -34,11 +34,11 @@ class RPNProposal(gluon.HybridBlock):
         These values must be the same as stds used in RPNTargetGenerator.
     """
     def __init__(self, clip, nms_thresh, train_pre_nms, train_post_nms,
-                 test_pre_nms, test_post_nms, min_size, stds):
+                 test_pre_nms, test_post_nms, min_size, stds, dtype='float32'):
         super(RPNProposal, self).__init__()
         self._box_to_center = BBoxCornerToCenter()
         self._box_decoder = NormalizedBoxCenterDecoder(stds=stds, clip=clip)
-        self._clipper = BBoxClipToImage()
+        self._clipper = BBoxClipToImage(dtype=dtype)
         # self._compute_area = BBoxArea()
         self._nms_thresh = nms_thresh
         self._train_pre_nms = max(1, train_pre_nms)
@@ -96,7 +96,6 @@ class RPNProposal(gluon.HybridBlock):
             pre = F.concat(score, roi, dim=-1)
             tmp = F.contrib.box_nms(pre, overlap_thresh=self._nms_thresh, topk=pre_nms,
                                     coord_start=1, score_index=0, id_index=-1, force_suppress=True)
-
             # slice post_nms number of boxes
             result = F.slice_axis(tmp, axis=1, begin=0, end=post_nms)
             rpn_scores = F.slice_axis(result, axis=-1, begin=0, end=1)
